+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 9.1.0 IN PROGRESS
 
 * Add STATUS FILTER to LOST ITEMS REQUIRING ACTUAL COST processing page. Refs UIU-2748.
+* Add "Lost items requiring actual cost" to Actions dropdown in User details record. Refs UIU-2810.
 
 ## [9.0.0](https://github.com/folio-org/ui-users/tree/v9.0.0) (2023-02-20)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v8.1.0...v9.0.0)